Updated loadsim with zprof and slc_prj
